+++ R/pkgaudit.R
@@ -75,7 +75,7 @@ new_pkgaudit <- function(
 # object is built at build time and R/ files are collated alphabetically.
 .phase_columns <- c(
   "at_autoconf", "at_build", "at_check", "at_install_src", "at_install_bin",
-  "on_load", "on_attach", "on_unload", "on_detach"
+  "at_load", "at_attach", "at_unload", "at_detach"
 )