+++ tests/testthat/fixtures/patterns/system_sys/negative_4.R
@@ -1,0 +1,1 @@
+foo$eval_fork(quote(1))